パーサーを言語ごとに分離した
